feat: add authentication to benchmark and bugfixing
