+++ Cargo.toml
@@ -6,11 +6,12 @@ actix-web = "4.2.1"
 tokio = { version = "1", features = ["full"] }
 sqlx = { version = "0.6", features = [ "runtime-actix-native-tls" , "postgres", "uuid", "json", "chrono" ] }
 log = "0.4.16"
-env_logger = "0.9.0"
+env_logger = "0.10.0"
 config = "0.13"
 serde = "1"
 serde_json = "1"
-dcl-rpc = "2.3.0" 
+dcl-rpc = "2.3.2"
+dcl-crypto = { git = "https://github.com/decentraland/decentraland-crypto-rust.git", rev = "db4e107307012ad8bd7bf7f56974163be9aec0d4" } 
 
 [profile.release]
 strip = true
