+++ crates/benchmark/Cargo.toml
@@ -16,6 +16,9 @@ rand = "0.8.5"
 reqwest = { version = "0.11.17", features = ["json"] }
 serde = "1.0"
 clap = { version = "4.2.7", features = ["derive"] }
+tungstenite = "0.18"
+serde_json = "1.0"
+dcl-crypto = { workspace = true }
 
 [build-dependencies]
 prost-build = "0.11.5"
