+++ crates/benchmark/src/main.rs
@@ -1,15 +1,13 @@
 use std::{thread::sleep, time::Duration};
 
 use clap::Parser;
-use dcl_rpc::{
-    client::RpcClient,
-    transports::web_socket::{WebSocketClient, WebSocketTransport},
-};
 use env_logger::init as initialize_logger;
-use log::info;
+use log::{debug, info};
 use quests_benchmark::{
-    simulation::{TestClient, TestContext},
-    Args, Simulation,
+    args::Args,
+    client::handle_client,
+    quests_simulation::{TestClient, TestContext},
+    simulation::Simulation,
 };
 use tokio::time::Instant;
 
@@ -27,7 +25,7 @@ async fn main() {
     let mut rpc_clients = vec![];
 
     for i in 0..args.clients {
-        set.spawn(handle_client(args.rpc_host.clone()));
+        set.spawn(handle_client(args.clone()));
         if (i + 1) % args.parallel as usize == 0 {
             while let Some(res) = set.join_next().await {
                 match res.unwrap() {
@@ -38,8 +36,9 @@ async fn main() {
 
                         info!("Connected clients: {}", rpc_clients.len());
                     }
-                    Err(_) => {
-                        info!("Ending test as clients can't connect to server anymore");
+                    Err(e) => {
+                        debug!("Couldn't create client: {e:?}");
+                        info!("Ending test as clients can't connect to server");
                         return;
                     }
                 }
@@ -61,28 +60,10 @@ async fn main() {
         "\nSimulation > Started and will run for {} minutes...",
         args.duration
     );
-    let duration = Duration::from_secs(60 * args.duration as u64);
-    Simulation::run::<TestContext, TestClient>(&args, rpc_clients, duration).await;
+    Simulation::run::<TestContext, TestClient>(&args, rpc_clients).await;
     info!("\nSimulation > Completed");
 }
 
 pub fn mean(values: &[u128]) -> u128 {
     values.iter().sum::<u128>() / values.len() as u128
-}
-
-pub async fn handle_client(
-    host: String,
-) -> Result<(RpcClient<WebSocketTransport>, u128, u128), ()> {
-    let whole_connection = Instant::now();
-    let ws = WebSocketClient::connect(&host).await.map_err(|e| {
-        log::error!("Couldn't connect to ws: {e:?}");
-    })?;
-    let transport = WebSocketTransport::new(ws);
-
-    let client_connection = Instant::now();
-    let client = RpcClient::new(transport).await.unwrap();
-    let client_creation_elapsed = client_connection.elapsed().as_millis();
-    let whole_connection = whole_connection.elapsed().as_millis();
-
-    Ok((client, whole_connection, client_creation_elapsed))
 }
