+++ crates/server/Cargo.toml
@@ -13,6 +13,7 @@ tokio = { workspace = true }
 serde = { workspace = true }
 config = { workspace = true }
 log = { workspace = true }
+dcl-crypto = { workspace = true }
 actix-web-prom = "0.6.0"
 async-trait = "0.1.57"
 tracing = "0.1"
@@ -33,7 +34,6 @@ derive_more = "0.99.17"
 openssl = { version = "0.10.45", features = ["vendored"] }
 warp = "0.3.3"
 tungstenite = "0.19.0"
-dcl-crypto = "0.1.0"
 fastrand = "1.9.0"
 
 [dev-dependencies]
